LOGIN AND REGISTRATION OF AN AGENT,IMPORTING OF BRANCH EXCEL DATA,VIEWING AND FILTERING OF MESSAGES
An agent is now able to be registered and to login to the platform. All of the BRANCH messages data has been successfully imported to the client messages model. The client messages are sorted according to key words before they are shown to the agent. The agent can also filter the messages according to the date and time.